Extract out unittest for dcolor
